Move and refactored casparclient
