+++ src/casparcg/casparclient.js
